MAGE-1134: change AlgoliaHelper to become a facade for AlgoliaConnector Service and move the logic
