+++ Test/Integration/Config/MultiStoreConfigTest.php
@@ -87,9 +87,7 @@ class MultiStoreConfigTest extends MultiStoreTestCase
             $fixtureSecondStore->getCode()
         );
 
-        $this->algoliaHelper->setStoreId($fixtureSecondStore->getId());
         $this->indicesConfigurator->saveConfigurationToAlgolia($fixtureSecondStore->getId());
-        $this->algoliaHelper->waitLastTask();
 
         $this->algoliaHelper->setStoreId($defaultStore->getId());
         $defaultCategoryIndexSettings = $this->algoliaHelper->getSettings($this->indexPrefix . 'default_categories');
